initial commit after setup script
