--- _package.json
@@ -1,43 +1,0 @@
-{
-  "name": "<%= packageName %>",
-  "description": "<%= description %>",
-  "repository": "https://github.com/<%= githubOrg %>/<%= shortName %>.git",
-  <%= publishInfo %>,
-  "scripts": {
-    "lint": "npm run lint:eslint && npm run lint:lit",
-    "lint:eslint": "eslint . --ext .js,.html",
-    "lint:lit": "lit-analyzer <%= shortName %>.js demo test",
-    "start": "es-dev-server --app-index demo/index.html --node-resolve --dedupe --open --watch",
-    "test": "npm run lint && npm run test:headless && npm run test:diff",
-    "test:diff": "mocha ./**/*.visual-diff.js -t 40000",
-    "test:diff:golden": "mocha ./**/*.visual-diff.js -t 40000 --golden",
-    "test:diff:golden:commit": "commit-goldens",
-    "test:headless": "karma start",
-    "test:headless:watch": "karma start --auto-watch=true --single-run=false",
-    "test:sauce": "karma start karma.sauce.conf.js"
-  },
-  "author": "D2L Corporation",
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "@brightspace-ui/visual-diff": "^1",
-    "@open-wc/testing": "^2",
-    "@open-wc/testing-karma": "^3",
-    "babel-eslint": "^10",
-    "deepmerge": "^3",
-    "es-dev-server": "^1",
-    "eslint": "^6",
-    "eslint-config-brightspace": "^0.6.7",
-    "eslint-plugin-html": "^6",
-    "eslint-plugin-lit": "^1",
-    "eslint-plugin-sort-class-members": "^1",
-    "frau-ci": "^1",
-    "karma-sauce-launcher": "^2",
-    "lit-analyzer": "^1",
-    "puppeteer": "^1"
-  },
-  "dependencies": {
-    "@brightspace-ui/core": "^1",
-    "@webcomponents/webcomponentsjs": "^2",
-    "lit-element": "^2"
-  }
-}
