+++ demo/index.html
@@ -5,18 +5,18 @@
 		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 		<script type="module">
 			import '@brightspace-ui/core/components/demo/demo-page.js';
-			import '../<%= shortName %>.js';
+			import '../tag-picker.js';
 		</script>
-		<title><%= name %></title>
+		<title>d2l-labs-tag-picker</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
 		<meta http-equiv="X-UA-Compatible" content="ie=edge">
 		<meta charset="UTF-8">
 	</head>
 	<body>
-		<d2l-demo-page page-title="<%= name %>">
-			<h2><%= name %></h2>
+		<d2l-demo-page page-title="d2l-labs-tag-picker">
+			<h2>d2l-labs-tag-picker</h2>
 			<d2l-demo-snippet>
-				<<%= name %>></<%= name %>>
+				<d2l-labs-tag-picker></d2l-labs-tag-picker>
 			</d2l-demo-snippet>
 		</d2l-demo-page>
 	</body>
