+++ test/tag-picker.visual-diff.html
@@ -4,16 +4,16 @@
 		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 		<script type="module">
 			import '@brightspace-ui/core/components/typography/typography.js';
-			import '../<%= shortName %>.js';
+			import '../tag-picker.js';
 		</script>
-		<title><%= name %> visual-diff tests</title>
+		<title>d2l-labs-tag-picker visual-diff tests</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
 		<meta http-equiv="X-UA-Compatible" content="ie=edge">
 		<meta charset="UTF-8">
 	</head>
 	<body class="d2l-typography">
 		<div class="visual-diff">
-			<<%= name %> id="default"></<%= name %>>
+			<d2l-labs-tag-picker id="default"></d2l-labs-tag-picker>
 		</div>
 	</body>
 </html>
